refactor(cli): extract nats-serve sub-app to cli_nats.py (ADR-059 V9)
- Move _probe_socket_daemon to nats/config.py
- Extract nats_app + nats_serve_tts/stt to cli_nats.py
- Remove import sys from cli.py (no longer used)
- Fix test_vram_guard.py import path for _probe_socket_daemon
- Update .importlinter: add new cli_* modules to top tier, add peer-import ignores
- Fix pyrightconfig.json: exclude GPU-dep files breaking the gate since b84e90a
- Update file_exemptions.txt: cli.py → #129, add cli_dictate.py exemption
