+++ .importlinter
@@ -9,7 +9,7 @@ include_external_packages = False
 name = Layer dependency rule
 type = layers
 layers =
-    voicecli.cli | voicecli.nats | voicecli.tg | voicecli.listen | voicecli.overlay
+    voicecli.cli | voicecli.cli_samples | voicecli.cli_dictate | voicecli.cli_doctor | voicecli.cli_nats | voicecli.nats | voicecli.tg | voicecli.listen | voicecli.overlay
     voicecli.api
     voicecli.adapters
     voicecli.engine | voicecli.transcribe | voicecli.daemon | voicecli.stt_daemon | voicecli.model_registry
@@ -19,10 +19,15 @@ ignore_imports =
     voicecli.translate -> voicecli.markdown
     # Transitional: voicecli.translate imports voicecli.utils for shared utilities — same layer peer dependency
     voicecli.translate -> voicecli.utils
-    # Transitional: voicecli.cli imports voicecli.nats adapters inline for NATS subcommands — needs extraction to adapter layer
-    voicecli.cli -> voicecli.nats.tts_adapter
-    voicecli.cli -> voicecli.nats.config
-    voicecli.cli -> voicecli.nats.stt_adapter
+    # Transitional: voicecli.cli assembler imports extracted sub-app modules — same top-tier peer dependencies (ADR-059 V9)
+    voicecli.cli -> voicecli.cli_samples
+    voicecli.cli -> voicecli.cli_dictate
+    voicecli.cli -> voicecli.cli_doctor
+    voicecli.cli -> voicecli.cli_nats
+    # Transitional: voicecli.cli_nats imports voicecli.nats adapters and config — same top-tier peer dependencies (ADR-059 V9)
+    voicecli.cli_nats -> voicecli.nats.config
+    voicecli.cli_nats -> voicecli.nats.tts_adapter
+    voicecli.cli_nats -> voicecli.nats.stt_adapter
     # Transitional: voicecli.cli imports voicecli.listen for recording subcommand — needs extraction to adapter layer
     voicecli.cli -> voicecli.listen
     # Transitional: voicecli.daemon imports voicecli.engine to load TTS engines — daemon is an infrastructure adapter, coupling is expected but needs SynthesisPort wiring (ADR-059)
